--- a/src/Fantomas.Core/AstTransformer.fs
+++ b/src/Fantomas.Core/AstTransformer.fs
@@ -1,5 +1,6 @@
 module internal Fantomas.Core.AstTransformer
 
+open FSharp.Compiler.SyntaxTrivia
 open FSharp.Compiler.Text
 open FSharp.Compiler.Syntax
 open Fantomas.Core.AstExtensions
@@ -48,6 +49,7 @@ let mkSynValSig (astNode: SynValSig) (r: range) (children: TriviaNode array) =
 
 let mkNodeOption (t: FsAstType) (r: range option) : TriviaNode option = Option.map (mkNode t) r
 
+// TODO: do we still need this when we have a clear anchor in the keyword?
 // We only need the let/type keyword anchor if there are xml docs or attributes present that have space between itself and the let/type keyword
 let mkNodeForRangeAfterXmlAndAttributes
     (t: FsAstType)
@@ -810,6 +812,35 @@ and visitSynExpr (synExpr: SynExpr) : TriviaNode list =
 
     visit synExpr id
 
+and visitSynLeadingKeyword (lk: SynLeadingKeyword) : TriviaNode =
+    match lk with
+    | SynLeadingKeyword.Let _ -> mkNode SynLeadingKeyword_Let lk.Range
+    | SynLeadingKeyword.LetRec _ -> mkNode SynLeadingKeyword_LetRec lk.Range
+    | SynLeadingKeyword.And _ -> mkNode SynLeadingKeyword_And lk.Range
+    | SynLeadingKeyword.Use _ -> mkNode SynLeadingKeyword_Use lk.Range
+    | SynLeadingKeyword.UseRec _ -> mkNode SynLeadingKeyword_UseRec lk.Range
+    | SynLeadingKeyword.Extern _ -> mkNode SynLeadingKeyword_Extern lk.Range
+    | SynLeadingKeyword.Member _ -> mkNode SynLeadingKeyword_Member lk.Range
+    | SynLeadingKeyword.MemberVal _ -> mkNode SynLeadingKeyword_MemberVal lk.Range
+    | SynLeadingKeyword.Override _ -> mkNode SynLeadingKeyword_Override lk.Range
+    | SynLeadingKeyword.OverrideVal _ -> mkNode SynLeadingKeyword_OverrideVal lk.Range
+    | SynLeadingKeyword.Abstract _ -> mkNode SynLeadingKeyword_Abstract lk.Range
+    | SynLeadingKeyword.AbstractMember _ -> mkNode SynLeadingKeyword_AbstractMember lk.Range
+    | SynLeadingKeyword.StaticMember _ -> mkNode SynLeadingKeyword_StaticMember lk.Range
+    | SynLeadingKeyword.StaticMemberVal _ -> mkNode SynLeadingKeyword_StaticMemberVal lk.Range
+    | SynLeadingKeyword.StaticAbstract _ -> mkNode SynLeadingKeyword_StaticAbstract lk.Range
+    | SynLeadingKeyword.StaticAbstractMember _ -> mkNode SynLeadingKeyword_StaticAbstractMember lk.Range
+    | SynLeadingKeyword.StaticVal _ -> mkNode SynLeadingKeyword_StaticVal lk.Range
+    | SynLeadingKeyword.StaticLet _ -> mkNode SynLeadingKeyword_StaticLet lk.Range
+    | SynLeadingKeyword.StaticLetRec _ -> mkNode SynLeadingKeyword_StaticLetRec lk.Range
+    | SynLeadingKeyword.StaticDo _ -> mkNode SynLeadingKeyword_StaticDo lk.Range
+    | SynLeadingKeyword.Default _ -> mkNode SynLeadingKeyword_Default lk.Range
+    | SynLeadingKeyword.DefaultVal _ -> mkNode SynLeadingKeyword_DefaultVal lk.Range
+    | SynLeadingKeyword.Val _ -> mkNode SynLeadingKeyword_Val lk.Range
+    | SynLeadingKeyword.New _ -> mkNode SynLeadingKeyword_New lk.Range
+    | SynLeadingKeyword.Do _ -> mkNode SynLeadingKeyword_Do lk.Range
+    | SynLeadingKeyword.Synthetic _ -> mkNode SynLeadingKeyword_Synthetic lk.Range
+
 and visitSynInterpolatedStringPart (synInterpolatedStringPart: SynInterpolatedStringPart) : TriviaNode =
     match synInterpolatedStringPart with
     | SynInterpolatedStringPart.String (_, range) -> mkNode SynInterpolatedStringPart_String range
@@ -843,13 +874,9 @@ and visitAnonRecordTypeField (_: Ident, t: SynType) = visitSynType t
 
 and visitSynMemberSig (ms: SynMemberSig) : TriviaNode =
     match ms with
-    | SynMemberSig.Member (valSig, mf, range) ->
+    | SynMemberSig.Member (valSig, _mf, range) ->
         let valSigNode = visitSynValSig valSig
-
-        mkNodeWithChildren
-            SynMemberSig_Member
-            range
-            (sortChildren [| yield! visitSynMemberFlags mf; yield! valSigNode.Children |])
+        mkNodeWithChildren SynMemberSig_Member range (sortChildren valSigNode.Children)
     | SynMemberSig.Interface (typeName, range) ->
         mkNodeWithChildren SynMemberSig_Interface range [| visitSynType typeName |]
     | SynMemberSig.Inherit (typeName, range) ->
@@ -959,13 +986,9 @@ and visitSynMemberDefn (mbrDef: SynMemberDefn) : TriviaNode =
             (sortChildren [| yield visitSynType inheritType; yield! visitSynExpr inheritArgs |])
     | SynMemberDefn.LetBindings (bindings, _, _, range) ->
         mkNodeWithChildren SynMemberDefn_LetBindings range (sortChildren [| yield! List.map visitSynBinding bindings |])
-    | SynMemberDefn.AbstractSlot (valSig, memberFlags, range) ->
+    | SynMemberDefn.AbstractSlot (valSig, _memberFlags, range) ->
         let valSigNode = visitSynValSig valSig
-
-        mkNodeWithChildren
-            SynMemberDefn_AbstractSlot
-            range
-            (sortChildren [| yield! visitSynMemberFlags memberFlags; yield! valSigNode.Children |])
+        mkNodeWithChildren SynMemberDefn_AbstractSlot range (sortChildren valSigNode.Children)
     | SynMemberDefn.Interface (typ, withKeyword, members, range) ->
         let ms =
             match members with
@@ -982,12 +1005,24 @@ and visitSynMemberDefn (mbrDef: SynMemberDefn) : TriviaNode =
     | SynMemberDefn.ValField (fld, range) -> mkNodeWithChildren SynMemberDefn_ValField range [| visitSynField fld |]
     | SynMemberDefn.NestedType (typeDefn, _, range) ->
         mkNodeWithChildren SynMemberDefn_NestedType range [| visitSynTypeDefn typeDefn |]
-    | SynMemberDefn.AutoProperty (attrs, _, _, typeOpt, _, _, _, _, _, equalsRange, synExpr, withKeyword, _, range) ->
+    | SynMemberDefn.AutoProperty (attrs,
+                                  _,
+                                  _,
+                                  typeOpt,
+                                  _,
+                                  _,
+                                  _,
+                                  _,
+                                  _,
+                                  synExpr,
+                                  range,
+                                  { EqualsRange = equalsRange
+                                    WithKeyword = withKeyword }) ->
         mkNodeWithChildren
             SynMemberDefn_AutoProperty
             range
             (sortChildren
-                [| yield mkNode SynMemberDefn_AutoProperty_Equals equalsRange
+                [| yield! Option.toList (mkNodeOption SynMemberDefn_AutoProperty_Equals equalsRange)
                    yield! Option.toList (mkNodeOption SynMemberDefn_AutoProperty_With withKeyword)
                    yield! (visitSynAttributeLists attrs)
                    yield! (Option.map visitSynType >> Option.toList) typeOpt
@@ -1048,7 +1083,7 @@ and visitSynSimplePats (sp: SynSimplePats) : TriviaNode =
 
 and visitSynBinding (binding: SynBinding) : TriviaNode =
     match binding with
-    | SynBinding (_, kind, _, _, attrs, preXml, valData, headPat, returnInfo, expr, _range, _, trivia) ->
+    | SynBinding (_, kind, _, _, attrs, preXml, _valData, headPat, returnInfo, expr, _range, _, trivia) ->
         let t =
             match kind with
             | SynBindingKind.StandaloneExpression -> SynBindingKind_StandaloneExpression
@@ -1061,12 +1096,8 @@ and visitSynBinding (binding: SynBinding) : TriviaNode =
             | _ -> Some(visitSynPat headPat)
 
         let keywordNode =
-            let triviaType, range =
-                match trivia.ExternKeyword with
-                | Some _ -> SynBinding_Extern, trivia.ExternKeyword
-                | None -> SynBinding_Let, trivia.LetKeyword
-
-            mkNodeForRangeAfterXmlAndAttributes triviaType preXml attrs range
+            let kwNode = visitSynLeadingKeyword trivia.LeadingKeyword
+            mkNodeForRangeAfterXmlAndAttributes kwNode.Type preXml attrs (Some kwNode.Range)
 
         let expr = SourceParser.parseExpressionInSynBinding returnInfo expr
         let returnInfo = Option.map visitSynBindingReturnInfo returnInfo
@@ -1077,21 +1108,10 @@ and visitSynBinding (binding: SynBinding) : TriviaNode =
             (sortChildren
                 [| yield! visitSynAttributeLists attrs
                    yield! keywordNode
-                   yield! visitSynValData valData
                    yield! Option.toList headPatNodes
                    yield! Option.toList returnInfo
                    yield! Option.toList (mkNodeOption SynBinding_Equals trivia.EqualsRange)
                    yield! visitSynExpr expr |])
-
-and visitSynValData (svd: SynValData) : TriviaNode list =
-    match svd with
-    | SynValData (Some memberFlags, _, _) -> visitSynMemberFlags memberFlags
-    | _ -> []
-
-and visitSynMemberFlags (memberFlags: SynMemberFlags) : TriviaNode list =
-    [ yield! Option.toList (mkNodeOption SynMemberFlags_Static memberFlags.Trivia.StaticRange)
-      yield! Option.toList (mkNodeOption SynMemberFlags_Member memberFlags.Trivia.MemberRange)
-      yield! Option.toList (mkNodeOption SynMemberFlags_Abstract memberFlags.Trivia.AbstractRange) ]
 
 and visitSynValSig (svs: SynValSig) : TriviaNode =
     match svs with
@@ -1100,13 +1120,13 @@ and visitSynValSig (svs: SynValSig) : TriviaNode =
             svs
             range
             (sortChildren
-                [| yield! Option.toList (mkNodeOption SynValSig_Val trivia.ValKeyword)
+                [| yield visitSynLeadingKeyword trivia.LeadingKeyword
                    yield visitSynIdent ident
                    yield! (visitSynAttributeLists attrs)
                    yield! Option.toList (visitSynValTyparDecls explicitValDecls)
                    yield visitSynType synType
                    yield! visitOptSynExpr expr
-                   yield! Option.toList (mkNodeOption SynValSig_With trivia.ValKeyword) |])
+                   yield! Option.toList (mkNodeOption SynValSig_With trivia.WithKeyword) |])
 
 and visitSynValTyparDecls (valTypeDecl: SynValTyparDecls) : TriviaNode option =
     match valTypeDecl with
@@ -1410,7 +1430,7 @@ and visitSynEnumCase (sec: SynEnumCase) : TriviaNode =
 
 and visitSynField (sfield: SynField) : TriviaNode =
     match sfield with
-    | SynField (attrs, _, ident, typ, _, preXmlDoc, _, _range) ->
+    | SynField (attrs, _, ident, typ, _, preXmlDoc, _, _range, trivia) ->
         let innerNode =
             match ident with
             | None -> []
@@ -1423,7 +1443,8 @@ and visitSynField (sfield: SynField) : TriviaNode =
             SynField_
             sfield.FullRange
             (sortChildren
-                [| yield! (visitSynAttributeLists attrs)
+                [| yield! Option.toList (Option.map visitSynLeadingKeyword trivia.LeadingKeyword)
+                   yield! (visitSynAttributeLists attrs)
                    yield! innerNode
                    yield visitSynType typ |])