… CHANGELOG.md
… paket.dependencies
… paket.lock
… src/Fantomas.Core/Selection.fs
… src/Fantomas.Core/SourceParser.fs
… src/Fantomas.Core/SourceTransformer.fs
… src/Fantomas.Core/Trivia.fs
… src/Fantomas.Core/TriviaTypes.fs
… src/Fantomas.FCS/Fantomas.FCS.fsproj