@dawedawe a lot of good stuff.
`SynLeadingKeyword` allows us to generalize almost all `SynBinding` nodes, regardless of their parent.

This streamlines and simplifies a lot of things in CodePrinter.