@baronfel 2022-10-13
Any impact on perf/memory usage with this change, now that's it's all in the parser?